feat: add tail and noLast
